refactor(changeRequest): move source components
