--- apps/concept-catalog/components/form-fields/source-for-definition/source-for-definition.module.css
@@ -1,14 +1,0 @@
-.row {
-  display: flex;
-  flex-direction: row;
-  gap: 1em;
-  align-items: flex-end;
-}
-
-.currentValue {
-  composes: currentValue from '../form-fields.module.css';
-}
-
-.container {
-  composes: containerWithLanguage from '../form-fields.module.css';
-}
