+++ apps/concept-catalog/components/form-fields/source-section/existing-sources/existing-sources.module.css
@@ -9,7 +9,7 @@
 }
 
 .existingValue {
-  composes: currentValue from '../form-fields.module.css';
+  composes: currentValue from '../../form-fields.module.css';
 }
 
 .sourceEntry {
