+++ apps/concept-catalog/components/form-fields/source-section/source-for-definition/source-for-definition.module.css
@@ -1,0 +1,14 @@
+.row {
+  display: flex;
+  flex-direction: row;
+  gap: 1em;
+  align-items: flex-end;
+}
+
+.currentValue {
+  composes: currentValue from '../../form-fields.module.css';
+}
+
+.container {
+  composes: containerWithLanguage from '../../form-fields.module.css';
+}
